fix(RAIN-38962) verify go source file names use _ rather than -
